# service-engine-x-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Service Engine X API",
    "description": "Service Engine X REST API - Multi-tenant SaaS platform for service businesses",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/engagements": {
      "get": {
        "tags": [
          "Engagements"
        ],
        "summary": "List Engagements",
        "description": "List all engagements for the authenticated organization.",
        "operationId": "list_engagements_api_engagements_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "created_at:desc",
              "title": "Sort"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 3,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status",
              "title": "Status"
            },
            "description": "Filter by status"
          },
          {
            "name": "client_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by client UUID",
              "title": "Client Id"
            },
            "description": "Filter by client UUID"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Engagements Api Engagements Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Engagements"
        ],
        "summary": "Create Engagement",
        "description": "Create a new engagement.",
        "operationId": "create_engagement_api_engagements_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EngagementCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngagementResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/engagements/{engagement_id}": {
      "get": {
        "tags": [
          "Engagements"
        ],
        "summary": "Retrieve Engagement",
        "description": "Retrieve an engagement with its projects and conversations.",
        "operationId": "retrieve_engagement_api_engagements__engagement_id__get",
        "parameters": [
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Engagement Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngagementResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Engagements"
        ],
        "summary": "Update Engagement",
        "description": "Update an engagement.",
        "operationId": "update_engagement_api_engagements__engagement_id__put",
        "parameters": [
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Engagement Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EngagementUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngagementResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Engagements"
        ],
        "summary": "Delete Engagement",
        "description": "Close an engagement (sets status to Closed).\n\nEngagements are not hard-deleted to preserve history.\nUse this to close an engagement relationship.",
        "operationId": "delete_engagement_api_engagements__engagement_id__delete",
        "parameters": [
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Engagement Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClientSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email"
        ],
        "title": "ClientSummary",
        "description": "Minimal client info for engagement responses."
      },
...
      "ConversationSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "status_id": {
            "type": "integer",
            "title": "Status Id"
          },
          "last_message_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Message At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "subject",
          "status",
          "status_id",
          "last_message_at"
        ],
        "title": "ConversationSummary",
        "description": "Minimal conversation info for engagement responses."
      },
...
      "EngagementCreate": {
        "properties": {
          "client_id": {
            "type": "string",
            "title": "Client Id",
            "description": "UUID of the client"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Optional display name"
          },
          "proposal_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposal Id",
            "description": "Originating proposal UUID"
          }
        },
        "type": "object",
        "required": [
          "client_id"
        ],
        "title": "EngagementCreate",
        "description": "Schema for creating an engagement."
      },
      "EngagementResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "org_id": {
            "type": "string",
            "title": "Org Id"
          },
          "client_id": {
            "type": "string",
            "title": "Client Id"
          },
          "client": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClientSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "status_id": {
            "type": "integer",
            "title": "Status Id"
          },
          "proposal_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposal Id"
          },
          "projects": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProjectSummary"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projects"
          },
          "conversations": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ConversationSummary"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversations"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "org_id",
          "client_id",
          "name",
          "status",
          "status_id",
          "proposal_id",
          "created_at",
          "updated_at",
          "closed_at"
        ],
        "title": "EngagementResponse",
        "description": "Schema for engagement response."
      },
...
      "EngagementUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "status": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 3.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "description": "1=Active, 2=Paused, 3=Closed"
          }
        },
        "type": "object",
        "title": "EngagementUpdate",
        "description": "Schema for updating an engagement."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "status_id": {
            "type": "integer",
            "title": "Status Id"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "phase_id": {
            "type": "integer",
            "title": "Phase Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "status_id",
          "phase",
          "phase_id"
        ],
        "title": "ProjectSummary",
        "description": "Minimal project info for engagement responses."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Token",
        "description": "API token authentication. Get your token from the dashboard."
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}